move current_repodata generator to its own file
