+++ conda_index/utils.py
@@ -57,3 +57,29 @@ def file_contents_match(pathA, pathB):
     """
 
     return filecmp.cmp(pathA, pathB, shallow=False)
+
+
+def human_bytes(n):
+    """
+    Return the number of bytes n in more human readable form.
+
+    Examples:
+        >>> human_bytes(42)
+        '42 B'
+        >>> human_bytes(1042)
+        '1 KB'
+        >>> human_bytes(10004242)
+        '9.5 MB'
+        >>> human_bytes(100000004242)
+        '93.13 GB'
+    """
+    if n < 1024:
+        return "%d B" % n
+    k = n / 1024
+    if k < 1024:
+        return "%d KB" % round(k)
+    m = k / 1024
+    if m < 1024:
+        return f"{m:.1f} MB"
+    g = m / 1024
+    return f"{g:.2f} GB"
